+++ lib/enums/timer_state.dart
@@ -1,0 +1,1 @@
+enum TimerState { none, running, paused, finished }
